refactor: move dumbo, mmds and logger to vmm (5/11)
Update MMDS to be a module.

Note: This is part of commits having the same title.
